+++ src/main/index.ts
@@ -1,14 +1,3 @@
-export {all, AllOptions} from './all';
-export {char} from './char';
-export {end} from './end';
-export {maybe} from './maybe';
-export {never} from './never';
-export {none} from './none';
-export {or} from './or';
-export {regex} from './regex';
-export {seq} from './seq';
-export {text, TextOptions} from './text';
-export {until, UntilOptions} from './until';
-export * from './taker-types';
+export * from './takers';
 export * from './token-utils';
 export * from './Tokenizer';
